+++ Rema1000/Rema1000/Models/Product.cs
@@ -13,9 +13,9 @@ namespace Eksamensprojekt_Rema1000.Models
         public int ProductId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Kilograms { get; set; }
+        public double Kilograms { get; set; }
         public int QuantityInPackage { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public int Stock { get; set; }
